Project/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
alembic==1.8.0
Babel==2.10.1
black==22.3.0
click==8.1.3
colorama==0.4.4
ecdsa==0.17.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.37
tomli==2.0.1
Werkzeug==2.1.2
wrapt==1.11.1
WTForms==3.0.1
